backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
feedparser==6.0.11
httpx==0.27.0
beautifulsoup4==4.12.3
geopy==2.4.1
apscheduler==3.10.4
pydantic==2.7.1
python-dateutil==2.9.0
lxml==5.2.2
aiofiles==23.2.1
